Move bluetooth device list to BluetoothSender.kt; Corrected permission requests locations and behaviour when permission is denied.
