+++ app/src/main/kotlin/de/markusfisch/android/binaryeye/activity/CameraActivity.kt
@@ -664,7 +664,7 @@ fun showResult(
 			}
 		}
 	}
-	if (prefs.sendScanBluetooth && prefs.sendScanBluetoothHost.isNotEmpty()) {
+	if (prefs.sendScanBluetooth && prefs.sendScanBluetoothHost.isNotEmpty() && activity.hasBluetoothPermission()) {
 		scan.sendBluetoothAsync(
 			prefs.sendScanBluetoothHost
 		) { con, send ->
